requirements.txt
streamlit==1.38.0
google-generativeai==0.8.3
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
